Add failing test for Unique Name
